requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
attrs
blinker
cachetools
certifi
charset-normalizer
click
duckduckgo-search>=4.1.1
filetype
frozenlist
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
importlib-metadata
Jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-core
langchain-google-genai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-sdk
langsmith
markdown-it-py
MarkupSafe
mdurl
msgpack
multidict
rank-bm25
narwhals
numpy
orjson
packaging
pandas
Pillow
propcache
proto-plus
protobuf
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic_core
pydeck
Pygments
Pympler
pyparsing
python-dateutil
python-dotenv
pytz
pytz-deprecation-shim
PyYAML
referencing
requests
requests-toolbelt
rich
rpds-py
rsa
six
smmap
sniffio
SQLAlchemy
streamlit
tenacity
toml
tornado
tqdm
typing_extensions
tzdata
tzlocal
uritemplate
urllib3
validators
yarl
zipp
langchain-community
langchain-core
langchain_openai
faiss-cpu
faker